+++ src/hooks/search/useSearchSortedResults.ts
@@ -1,9 +1,9 @@
 import { useInfiniteQuery } from '@tanstack/react-query';
 import { useEffect } from 'react';
 
+import { filterKeywordSortedDrinks } from '@/app/actions/filter';
 import useSearchStore from '@/store/keywordStore';
 import useSortStore from '@/store/selectStore';
-import { filterKeywordSortedDrinks } from '@/utils/filter/action';
 
 const useSearchSortedResults = () => {
   const { keyword, searchTriggerFetch, setSearchTriggerFetch } =
